Rename Lines => rows
Makes more sense with the content inside it
