+++ app/Parser.tsx
@@ -11,9 +11,9 @@ import { JSONTree } from "react-json-tree";
 import { lexer, parse, refineRowType, splitToCleanRows } from "./parse";
 import * as Ariakit from "@ariakit/react";
 import { ErrorBoundary } from "react-error-boundary";
-import { TreeRow } from "./Lines/TreeRow";
-import { ClientReferenceRow } from "./Lines/ClientReferenceRow";
-import { HintRow } from "./Lines/HintRow";
+import { TreeRow } from "./rows/TreeRow";
+import { ClientReferenceRow } from "./rows/ClientReferenceRow";
+import { HintRow } from "./rows/HintRow";
 import { GenericErrorBoundaryFallback } from "./GenericErrorBoundaryFallback";
 import { stringToKiloBytes } from "./stringtoKiloBytes";
 import { PayloadContext } from "./PayloadContext";
