+++ app/rows/HintRow.tsx
